requirements.txt
pandas
datasets
langchain-groq
python-dotenv
langchain_community
pyttsx3
plotly
matplotlib
wordcloud